Fix graphite vulkan build errors (define GetProc famility of functions where it can be used by both, ganesh and graphite)
