+++ skia-safe/Cargo.toml
@@ -18,7 +18,7 @@ categories = [
 ]
 license = "MIT"
 
-version = "0.153.1"
+version = "0.153.2"
 authors = ["Armin Sander <armin@replicator.org>"]
 edition.workspace = true
 rust-version.workspace = true
@@ -63,7 +63,7 @@ graphite = ["skia-bindings/graphite"]
 
 [dependencies]
 bitflags = "2.0"
-skia-bindings = { version = "=0.153.1", path = "../skia-bindings", default-features = false }
+skia-bindings = { version = "=0.153.2", path = "../skia-bindings", default-features = false }
 
 
 # D3D types & ComPtr
