+++ skia-safe/src/gpu/vk/vulkan_types.rs
@@ -1,8 +1,37 @@
+use std::{ffi::CStr, os::raw};
+
 use skia_bindings::{
     self as sb, skgpu_VulkanAlloc, skgpu_VulkanBackendMemory, skgpu_VulkanYcbcrConversionInfo,
 };
 
 use crate::{gpu::vk, prelude::*};
+
+// TODO: Tried to use CStr here, but &CStr needs a lifetime parameter
+//       which would make the whole GetProc trait generic.
+#[derive(Copy, Clone, Debug)]
+pub enum GetProcOf {
+    Instance(vk::Instance, *const raw::c_char),
+    Device(vk::Device, *const raw::c_char),
+}
+
+impl GetProcOf {
+    /// # Safety
+    /// The referred raw `name` strings must outlive the returned CStr reference.
+    pub unsafe fn name(&self) -> &CStr {
+        match *self {
+            GetProcOf::Instance(_, name) => unsafe { CStr::from_ptr(name) },
+            GetProcOf::Device(_, name) => unsafe { CStr::from_ptr(name) },
+        }
+    }
+}
+
+// TODO: Really would like to see a fn() signature here, but I'm always running
+//       into a conflict between extern "C" and extern "system".
+pub type GetProcResult = *const raw::c_void;
+
+// GetProc is a trait alias for Fn...
+pub trait GetProc: Fn(GetProcOf) -> GetProcResult {}
+impl<T> GetProc for T where T: Fn(GetProcOf) -> GetProcResult {}
 
 #[deprecated(since = "0.76.0", note = "Use BackendMemory")]
 pub type GraphicsBackendMemory = skgpu_VulkanBackendMemory;
